Rename web/lib/schema to web/lib/validation-schema
