+++ apps/web/components/admin-notes/AdminNoteView.tsx
@@ -14,7 +14,7 @@ import {
   DialogTrigger,
 } from '../ui/dialog';
 import { AdminNoteRouteTargetEnumHelper } from '@/lib/enum-helpers';
-import { adminNoteFormSchema } from '@/lib/schema';
+import { adminNoteFormSchema } from '@/lib/validation-schema';
 import { hasAdminScope } from '@/lib/auth/roles';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
