+++ apps/web/components/beatmaps/list/BeatmapListFilter.tsx
@@ -11,7 +11,7 @@ import { Ruleset } from '@otr/core/osu';
 import {
   beatmapListFilterSchema,
   defaultBeatmapListFilter,
-} from '@/lib/schema';
+} from '@/lib/validation-schema';
 import { RulesetEnumHelper } from '@/lib/enum-helpers';
 import {
   Form,
