+++ apps/web/components/leaderboard/LeaderboardFilter.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useMemo, useState } from 'react';
-import { leaderboardFilterSchema } from '@/lib/schema';
+import { leaderboardFilterSchema } from '@/lib/validation-schema';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
