+++ apps/web/components/scores/ScoreAdminView.tsx
@@ -10,7 +10,7 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
-import { scoreEditFormSchema } from '@/lib/schema';
+import { scoreEditFormSchema } from '@/lib/validation-schema';
 import { cn } from '@/lib/utils';
 import { zodResolver } from '@hookform/resolvers/zod';
 import {
