+++ apps/web/components/tournaments/list/TournamentListFilter.tsx
@@ -11,7 +11,7 @@ import { Input } from '@/components/ui/input';
 import {
   defaultTournamentListFilter,
   tournamentListFilterSchema,
-} from '@/lib/schema';
+} from '@/lib/validation-schema';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Search, ArrowUp, ArrowDown, Filter, X, Calendar } from 'lucide-react';
